coloca um componente na tela de posts por hashtag e altera um pouco a funçncao do componente post
